TestLandscapes.jl

This is a collection of reference energy landscapes that frequently appear in the literature.

TestLandscapes.jl

1D Landscapes

2D Landscapes

Index

1D Landscapes

TestLandscapes.SymmetricDoubleWell — Method

SymmetricDoubleWell - The classic symmetric doublewell potential.

Fields

x - Position x

TestLandscapes.AsymmetricDoubleWell — Method

AsymmetricDoubleWell - An asymmetric double well potential requires |δ|< 1 for multiple minima.

Fields

x - Position x

Optional Fields

δ = 0.5 - Asymmetry parameter

TestLandscapes.FatSkinnyDoubleWell10 — Method

FatSkinnyDoubleWell10 - The fat-skinny double well of degree 10. This introduces entropic effects in dimension 1.

Fields

x - Position x

2D Landscapes

TestLandscapes.EntropicSwitch — Method

EntropicSwitch - Entropically switching potential with three local minima. It is symmetric about x=0

Fields

x - Position x in R²

TestLandscapes.SymmetricTwoChannel — Method

SymmetricTwoChannel - Double well potential in 2D with two, symmetric channels joining them.

Fields

x - Position x in R²

TestLandscapes.Muller — Method

Muller - The Muller potential with three distinct minima and highy asymmetric.

Fields

x - Position x in R²

TestLandscapes.Rosenbrock — Method

Rosenbrock - Banana shaped Rosenbrock potentials with global minimum is located at (a,a²).

Fields

x - Position x in R²

Optional Fields

a = 1.0 - Rosenbrock parameter

b = 100.0 - Rosenbrock parameter

TestLandscapes.Zpotential — Method

Zpotential - Z shaped potential.

Fields

x - Position x in R²

TestLandscapes.EntropicBox — Method

EntropicBox - A potential concentrated in [0,1]² with internal entropic barriers. Formulated by D. Aristoff (Colorado State).

Fields

x - Position x in R²

Index

TestLandscapes.AsymmetricDoubleWell

TestLandscapes.EntropicBox

TestLandscapes.EntropicSwitch

TestLandscapes.FatSkinnyDoubleWell10

TestLandscapes.Muller

TestLandscapes.Rosenbrock

TestLandscapes.SymmetricDoubleWell

TestLandscapes.SymmetricTwoChannel

TestLandscapes.Zpotential