2D Landscapes

TestLandscapes.EntropicSwitch — Method

EntropicSwitch - Entropically switching potential with three local minima. It is symmetric about x=0

Fields

x - Position x in R²

TestLandscapes.SymmetricTwoChannel — Method

SymmetricTwoChannel - Double well potential in 2D with two, symmetric channels joining them.

Fields

x - Position x in R²

TestLandscapes.Muller — Method

Muller - The Muller potential with three distinct minima and highy asymmetric.

Fields

x - Position x in R²

TestLandscapes.Rosenbrock — Method

Rosenbrock - Banana shaped Rosenbrock potentials with global minimum is located at (a,a²).

Fields

x - Position x in R²

Optional Fields

a = 1.0 - Rosenbrock parameter

b = 100.0 - Rosenbrock parameter

TestLandscapes.Zpotential — Method

Zpotential - Z shaped potential.

Fields

x - Position x in R²

TestLandscapes.EntropicBox — Method

EntropicBox - A potential concentrated in [0,1]² with internal entropic barriers. Formulated by D. Aristoff (Colorado State). This was used in https://pubs.aip.org/aip/jcp/article/158/1/014108/2867485

Fields

x - Position x in R²